Add dup, qos, retain params to MQTT on_message event (#172)
Also, payload is now `Uint8Array`

Previously, its type claimed to be `ArrayBuffer` in one callback and `Buffer` in another callback. In fact it was **always** `ArrayBuffer` in the Native impl and **always** `Buffer` in the Browser impl. We needed to settle on a common type to pass the user. Chose `Uint8Array` because it's a native JS type that doesn't require a copy of the underlying data when converting from `Buffer` or `ArrayBuffer`.