Fix "dubious ownership" error from git, when running in docker using bind mounted directory (#365)
**Issue:**
This docker image uses a newer version of git. When running with a bind mounted directory, git fails with "dubious ownership" errors.

Newer versions of git fail if the current user doesn't match the owner of the `.git/` folder. Since the directory is bind mounted, it is owned by the user on the host OS. But within a container, the user is `root` by default. Since `root` doesn't match the owner of the `.git/` folder, git fails.

**Description of changes:**
Set `--user` to match the owner of the bind mounted directory.

See the original diagnosis of the issue here: https://github.com/awslabs/aws-crt-python/pull/404