Fix "Cannot find module" for aws-crt-nodejs
In lib/native/bindings.js, after successfully finding the aws-crt-nodejs native module, it drops the *.node* extension when calling require.
That should be fine in a pure commonJs environment but dropping the extension breaks some ES-like runtime and build environments.

This PR includes the extension in the require call.

I don't see a test to update. (Consider adding some? There seem to be a lot of regressions in aws-sdk-js-v3 relating to aws-crt and signature-v4-crt. It would take a lot of work around test fixtures but perhaps worthwhile/justified?)